Add initial support for a C++11 compiled protobuf
This adds the initial suppport for the C++11/libc++ compiled version of
protobuf which can be used with projects that require a libc++ version.

The biggst issue is that protobuf uses strings internally, and those can
not be shared across different versions of the library (pre 2011) due to
implementation differences.